feat(profile): wire Follow + add Endorse to the mobile header
The mobile <ProfileHeader> Follow button was a dead placeholder and had
no Endorse action, while the desktop sidebar carried the full flow.

Extract the sidebar's FollowButton + EndorseButton into a shared
`profile-follow-endorse` module plus a `FollowEndorseActions` wrapper, and
render it in both surfaces. Mobile now shows a working Follow with the
Endorse button to its right (follow/unfollow + endorse/revoke, both
acting-as-group aware) — one implementation, no divergence.
