+++ src/components/profile/profile-sidebar.tsx
@@ -19,7 +19,6 @@ import {
 } from "lucide-react"
 import Avatar from "@/components/ui/avatar"
 import Button from "@/components/ui/button"
-import ConfirmDialog from "@/components/ui/confirm-dialog"
 import Input from "@/components/ui/input"
 import LoadingSpinner from "@/components/ui/loading-spinner"
 import SmartLink from "@/components/ui/smart-link"
@@ -34,24 +33,11 @@ import type { Group } from "@/lib/groups/types"
 import { useFollowing } from "@/hooks/use-following"
 import { useFollowers } from "@/hooks/use-followers"
 import { formatGraphCount } from "@/lib/utils/format-graph-count"
-import { useGivenEndorsements } from "@/hooks/use-endorsements"
+import { useReceivedEndorsements } from "@/hooks/use-received-endorsements"
 import {
-  useReceivedEndorsements,
-  addOptimisticReceivedEndorsement,
-  removeOptimisticReceivedEndorsement,
-} from "@/hooks/use-received-endorsements"
-import { useEndorsementLists } from "@/hooks/use-endorsement-lists"
-import { useAuthorInfo } from "@/hooks/use-author-info"
-import EndorseReasonModal, {
-  type EndorseReasonActingAs,
-} from "@/components/profile/endorse-reason-modal"
-import { runEndorseReasonConfirm } from "@/components/profile/endorse-reason-confirm"
-import { createFollow, deleteFollow, listFollowing } from "@/lib/atproto/follow"
-import {
-  createEndorsementAward,
-  deleteEndorsementAward,
-} from "@/lib/atproto/badges"
-import { appendItemToList } from "@/lib/atproto/collection"
+  FollowButton,
+  EndorseButton,
+} from "@/components/profile/profile-follow-endorse"
 import AddToListMenu from "@/components/lists/add-to-list-menu"
 import { LIST_ACCOUNTS_TYPE } from "@/lib/atproto/typed-lists"
 import type { CertifiedProfile } from "@/lib/atproto/types"
@@ -720,302 +706,3 @@ function AvatarEditOverlay({ onFile, isUploading, hasPending }: AvatarEditOverla
   )
 }
 
-/* ------------------------------ Follow ------------------------------
- *
- * Sidebar Follow / Following toggle, shown to signed-in viewers who
- * are looking at someone else's profile. Hidden for own-profile views
- * and for viewers who can edit (the Edit button takes the slot).
- *
- * Optimistically flips its own label between "Follow" and "Following"
- * while the write is in flight; the parent's `onAfterWrite` re-pages
- * the viewer's follow list so the true state catches up.
- */
-
-interface FollowButtonProps {
-  viewerDid: string
-  subjectDid: string
-  /** When set (acting-as-group), the follow write routes to the group's
-   *  repo via the BFF instead of the viewer's personal PDS. */
-  targetDid?: string
-  isFollowing: boolean
-  isLoading: boolean
-  /** Fired after a successful createFollow with the new record's
-   *  strong ref. Caller does the optimistic state update on both
-   *  the viewer's "following" hook and the subject's "followers"
-   *  hook. */
-  onFollowed: (uri: string, cid: string) => void
-  /** Fired after a successful deleteFollow. Caller does the
-   *  optimistic state update mirroring `onFollowed`. */
-  onUnfollowed: () => void
-}
-
-function FollowButton({
-  viewerDid,
-  subjectDid,
-  targetDid,
-  isFollowing,
-  isLoading,
-  onFollowed,
-  onUnfollowed,
-}: FollowButtonProps) {
-  const [isWriting, setIsWriting] = useState(false)
-  const disabled = isLoading || isWriting
-  // The repo the follow record lives in — the group when delegating, the
-  // viewer otherwise. Drives both the rkey lookup (list the right repo)
-  // and which repo's follows we delete from.
-  const actingRepo = targetDid ?? viewerDid
-
-  const handleClick = async () => {
-    if (disabled) return
-    const next = !isFollowing
-    setIsWriting(true)
-    try {
-      if (next) {
-        const result = await createFollow(viewerDid, subjectDid, { targetDid })
-        // Hand the new ref to the parent so the acting repo's following
-        // set and the subject's follower list update instantly.
-        onFollowed(result.uri, result.cid)
-      } else {
-        // Unfollow path: walk the acting repo's follows to find the rkey
-        // targeting this subject. Fetched fresh here to handle the
-        // duplicate-follow edge case (delete the most recent record).
-        // When acting as a group, the records live in the group's repo.
-        const { records } = await listFollowing(actingRepo, undefined, {
-          noCache: true,
-        })
-        const match = records
-          .filter((r) => r.value.subject === subjectDid)
-          .sort((a, b) =>
-            a.value.createdAt < b.value.createdAt ? 1 : -1,
-          )[0]
-        if (match) {
-          await deleteFollow(viewerDid, match.rkey, { targetDid })
-        }
-        onUnfollowed()
-      }
-    } catch (err) {
-      console.error("Follow toggle failed:", err)
-    } finally {
-      setIsWriting(false)
-    }
-  }
-
-  return (
-    <Button
-      variant={isFollowing ? "secondary" : "primary"}
-      size="sm"
-      onClick={handleClick}
-      disabled={disabled}
-      aria-pressed={isFollowing}
-    >
-      {isFollowing ? (
-        <Check size={14} strokeWidth={1.75} aria-hidden />
-      ) : (
-        <UserPlus size={14} strokeWidth={1.75} aria-hidden />
-      )}
-      {isFollowing ? "Following" : "Follow"}
-    </Button>
-  )
-}
-
-/**
- * Sidebar count formatter. `null` (not yet loaded) renders as an em
- * dash so the layout doesn't jump; `0` renders as `0` because that's
- * a real, meaningful value to viewers ("nobody follows this account
- * yet").
- */
-/* `formatGraphCount` now lives in @/lib/utils/format-graph-count and is shared
-   with the mobile <ProfileHeader> so both surfaces render identical counts. */
-
-/* ----------------------------- Endorse ------------------------------
- *
- * Endorse / Endorsed toggle. Lives next to FollowButton on foreign
- * profiles. Writes against the viewer's default endorsement
- * definition (`ensureEndorsementDefinition` runs implicitly inside
- * `createEndorsementAward`). Revoking gates on a Confirm dialog
- * because endorsement deletion is silent on the recipient's side.
- *
- * Used to live as `EndorseShortcut` at the top of the Received
- * sub-tab — moved into the sidebar so the action sits with Follow,
- * and discovery doesn't depend on the viewer landing on the
- * Endorsements tab first.
- */
-
-interface EndorseButtonProps {
-  viewerDid: string
-  subjectDid: string
-  /** The DID the endorsement is authored AS — the active group when
-   *  delegating, the viewer otherwise. Drives the given-endorsements
-   *  set that decides Endorse / Endorsed and the revoke rkey lookup. */
-  actingDid: string
-  /** The active group, when the viewer is acting as one. `null` for the
-   *  personal path. Used to route the write to the group's repo and to
-   *  name the parties in the reason modal. */
-  activeOrg: Group | null
-  /** The signed-in operator's handle, surfaced in the delegation header
-   *  of the reason modal. */
-  operatorHandle: string | null
-}
-
-function EndorseButton({
-  viewerDid,
-  subjectDid,
-  actingDid,
-  activeOrg,
-  operatorHandle,
-}: EndorseButtonProps) {
-  // Acting-as-group flag + the group DID the write routes to.
-  const targetDid = activeOrg?.groupDid
-  const ownGiven = useGivenEndorsements(actingDid)
-  // Endorsement lists are personal-only — never load or surface them
-  // while acting as a group (the reason modal hides the picker too).
-  const ownLists = useEndorsementLists(activeOrg ? null : viewerDid)
-  const { info: subjectInfo } = useAuthorInfo(subjectDid)
-  const subjectLabel =
-    subjectInfo?.displayName || subjectInfo?.handle || subjectDid
-  const [isWriting, setIsWriting] = useState(false)
-  const [confirmRevoke, setConfirmRevoke] = useState(false)
-  const [reasonOpen, setReasonOpen] = useState(false)
-  // Optimistic flip — mirrors FollowButton. The hook's `endorsements`
-  // refetch may lag the PDS write by a beat, so we override locally
-  // until the parent value catches up.
-  const [optimistic, setOptimistic] = useState<boolean | null>(null)
-
-  const existing = ownGiven.endorsements.find((e) => e.subjectDid === subjectDid)
-  const isEndorsedFromState = !!existing
-  const isEndorsed = optimistic ?? isEndorsedFromState
-
-  // Once the parent confirms the same state, drop the override so
-  // the hook value becomes authoritative again.
-  useEffect(() => {
-    if (optimistic !== null && isEndorsedFromState === optimistic) {
-      setOptimistic(null)
-    }
-  }, [isEndorsedFromState, optimistic])
-
-  const disabled = isWriting || ownGiven.isLoading
-
-  /**
-   * Click on the unfilled button → open the reason modal. The actual
-   * write happens inside the modal's confirm so the note travels
-   * with the award. If the user cancels the modal we never touch
-   * the optimistic state — the button stays on "Endorse".
-   */
-  const handleEndorseClick = () => {
-    if (disabled) return
-    setReasonOpen(true)
-  }
-
-  const handleReasonConfirm = async (note: string, listRkey: string | null) => {
-    setOptimistic(true)
-    setIsWriting(true)
-    try {
-      // quality-048: orchestration (create → refetch-given → optional
-      // list-append) lives in `runEndorseReasonConfirm` so the ordering
-      // and its error handling are unit-testable. It refetches the
-      // given set BEFORE the list-append and, on append failure, keeps
-      // optimistic=true while rethrowing the attribution error.
-      await runEndorseReasonConfirm({
-        note,
-        listRkey,
-        createAward: (n) =>
-          createEndorsementAward(viewerDid, subjectDid, n, { targetDid }),
-        // PR #110: as soon as the award lands, push it into the shared
-        // received-endorsements overlay so the subject's "Endorsed by N"
-        // counter and the Endorsements tab reflect it immediately, ahead
-        // of the 5-min scan cache / indexer catching up. The issuer is
-        // the acting repo (the group when delegating).
-        onAwardCreated: (award) =>
-          addOptimisticReceivedEndorsement(subjectDid, {
-            uri: award.uri,
-            cid: award.cid,
-            issuerDid: actingDid,
-            createdAt: new Date().toISOString(),
-            note: note || undefined,
-            responseState: null,
-          }),
-        appendToList: (rkey, award) => appendItemToList(viewerDid, rkey, award),
-        refetchGiven: () => ownGiven.refetch(),
-        refetchLists: () => ownLists.refetch(),
-        setOptimistic,
-      })
-      setReasonOpen(false)
-    } catch (err) {
-      // The optimistic flag is managed inside the orchestrator: cleared
-      // on an award failure, kept on a list-append failure. Just rethrow
-      // so the modal surfaces the error and stays open.
-      throw err
-    } finally {
-      setIsWriting(false)
-    }
-  }
-
-  const handleConfirmRevoke = async () => {
-    if (!existing || disabled) return
-    setOptimistic(false)
-    setIsWriting(true)
-    try {
-      await deleteEndorsementAward(viewerDid, existing.rkey, { targetDid })
-      removeOptimisticReceivedEndorsement(subjectDid, existing.uri)
-      await ownGiven.refetch()
-      setConfirmRevoke(false)
-    } catch (err) {
-      console.error("Revoke endorsement failed:", err)
-      setOptimistic(null)
-    } finally {
-      setIsWriting(false)
-    }
-  }
-
-  // Delegation naming for the reason modal — names the group, the
-  // operator, and the subject when acting as a group. `null` on the
-  // personal path leaves the modal in its default single-party shape.
-  const actingAs: EndorseReasonActingAs | undefined = activeOrg
-    ? {
-        orgName: activeOrg.displayName || activeOrg.handle,
-        orgHandle: activeOrg.handle,
-        operatorHandle: operatorHandle ?? "you",
-        operatorRole: activeOrg.role,
-      }
-    : undefined
-
-  return (
-    <>
-      <Button
-        variant={isEndorsed ? "secondary" : "primary"}
-        size="sm"
-        onClick={isEndorsed ? () => setConfirmRevoke(true) : handleEndorseClick}
-        disabled={disabled}
-        aria-pressed={isEndorsed}
-      >
-        {isEndorsed ? (
-          <Check size={14} strokeWidth={1.75} aria-hidden />
-        ) : (
-          <ThumbsUp size={14} strokeWidth={1.75} aria-hidden />
-        )}
-        {isEndorsed ? "Endorsed" : "Endorse"}
-      </Button>
-      {reasonOpen ? (
-        <EndorseReasonModal
-          subjectLabel={subjectLabel}
-          actingAs={actingAs}
-          lists={ownLists.lists.map((l) => ({ rkey: l.rkey, title: l.title }))}
-          onConfirm={handleReasonConfirm}
-          onClose={() => setReasonOpen(false)}
-        />
-      ) : null}
-      {confirmRevoke ? (
-        <ConfirmDialog
-          title="Revoke endorsement?"
-          message="Your endorsement will be removed from this profile. You can endorse them again later."
-          confirmLabel="Revoke"
-          cancelLabel="Keep endorsement"
-          confirmVariant="destructive"
-          isConfirming={isWriting}
-          onConfirm={handleConfirmRevoke}
-          onCancel={() => !isWriting && setConfirmRevoke(false)}
-        />
-      ) : null}
-    </>
-  )
-}
